+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for polysemy-composite
 
+## v0.1.3.0
+
+* Add `separateRecInitial`, `stripRecInitial`, `stripRecInitial`, `endRecInitial`.
+* Add `separateRecTerminal`, `stripRecTerminal`, `stripRecTerminal`, `endRecTerminal`.
+* Add `runRecInitialAsInputCompose`.
+
 ## v0.1.2.0
 
 * Inline fix.